changed: move EmbeddedArt into a separate file in utils
move it out of MUSIC_INFO namespace. for reuse
